refactor: trustpolicy
